+++ hilt/processor/src/main/kotlin/model/BindingBuilder.kt
@@ -58,7 +58,8 @@ class BindingBuilder(private val processingEnv: ProcessingEnvironment) {
             annotatedClassName = element.className(),
             supertypeClassName = (visitor.supertypeClassName ?: getSupertypeClassName(element)) as ClassName,
             componentClassName = visitor.componentClassName as? ClassName ?: defaultComponentClassName,
-            annotations = element.scopesAndQualifiers().map(AnnotationSpec::get)
+            annotations = element.scopesAndQualifiers().map(AnnotationSpec::get),
+            isTest = annotationType.simpleName.startsWith(TEST_ANNOTATION_PREFIX)
         )
     }
 
@@ -77,6 +78,8 @@ class BindingBuilder(private val processingEnv: ProcessingEnvironment) {
     }
 
     companion object {
+        private const val TEST_ANNOTATION_PREFIX = "Test"
+
         private val defaultComponentClassName = ClassName.get(SingletonComponent::class.java)
     }
 }